% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{green_ngs_palette}
\alias{green_ngs_palette}
\title{A Twist brand thematic color palette generator for ggplot
Function to extract Twist colors as hex codes}
\usage{
green_ngs_palette(...)
}
\arguments{
\item{...}{Character names of Twist colors for each product theme}
}
\description{
A Twist brand thematic color palette generator for ggplot
Function to extract Twist colors as hex codes
}
