% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{twist_color_pal}
\alias{twist_color_pal}
\title{Return function to interpolate a Twist color palette}
\usage{
twist_color_pal(palette = "green_ngs", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in twist_palettes for twist_color_pal}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a Twist color palette
}
