% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{twist_palettes}
\alias{twist_palettes}
\title{list of all Twist color palettes}
\format{
An object of class \code{list} of length 7.
}
\usage{
twist_palettes
}
\description{
Mother list of all Twist color palettes, generated by calling
individual color functions with their default parameters.
}
\section{Available palette names}{


"green_ngs", "grays", "blue_oligo_pools", "purple_genes",
"pink_biopharma", "orange_variant_library", "golds"
}

\keyword{datasets}
